pyproject.toml
[project]
name = "freqtrade-demo"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "freqtrade[freqai,hyperopt]",
    "nbformat>=4.2.0",
    "numpy>=2.0.0,<2.4.0",
    "scikit-learn>=1.8.0",
    "pyqlib>=0.9.0",
    "lightgbm>=4.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.18.0,<2.0.0",
    "pydmd>=2025.8.1",
    "alphalens-reloaded>=0.4.6",
    "empyrical-reloaded>=0.5.12",
]

[tool.uv]
link-mode = "copy"

[tool.uv.sources]
freqtrade = { git = "https://github.com/freqtrade/freqtrade.git", rev = "8e91fea11f3cc0354b952a5b667628b18378c7fe" }

[dependency-groups]
dev = [
    "ipykernel",
    "ipywidgets>=8.1.8",
    "plotly",
    "pytest>=9.0.2",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
]
